+++ src/mission/imap/+mag/+imap/+hk/Processor.m
@@ -2,6 +2,12 @@ classdef Processor < mag.HK
 % PROCESSOR Class containing MAG processor HK packet data.
 
     properties (Dependent)
+        % SINGLEBITERRORS Single bit errors in processor SRAM.
+        SingleBitErrors (:, 1) double
+        % MISSEDITFS Number of missed ITF frames.
+        MissedITFs (:, 1) double
+        % REJECTEDITFS Number of rejected ITF frames.
+        RejectedITFs (:, 1) double
         % FOBQUEUENUMMSG Outboard sensor message queue.
         FOBQueueNumMSG (:, 1) double
         % FIBQUEUENUMMSG Inboard sensor message queue.
@@ -9,6 +15,18 @@ classdef Processor < mag.HK
     end
 
     methods
+
+        function singleBitErrors = get.SingleBitErrors(this)
+            singleBitErrors = this.Data.SRAM_SINGBITERRCNT;
+        end
+
+        function missedITFs = get.MissedITFs(this)
+            missedITFs = this.Data.ITF_MSD_FR;
+        end
+
+        function rejectedITFs = get.RejectedITFs(this)
+            rejectedITFs = this.Data.ITF_REJ_FR;
+        end
 
         function fobQueueNumMSG = get.FOBQueueNumMSG(this)
             fobQueueNumMSG = this.Data.OBNQ_NUM_MSG;